fix: initialise codelines correctly
The codelines jquery object doesn't automatically create all the divs
it needs, it waits until you scroll them into view.
This is problematic because we want to be able to render breakpoints,
the current PC etc. and have them be present when you scroll to them
The solution is to manually create them.

Fixes #5