refactor: Create an InterpreterTrait to allow some abstraction
Refactor all the elements of Interpreter to not be public, and add a
trait with the required interface.
Rust doesn't have full inheritance, and we don't want the frontend to
have to know/care about backend types, so all data to and from the UI
is passed as strings and parsed in the backend.

This doesn't yet allow multiple backends but it's a step in the right direction