fix: Various instruction rendering issues
There were various bugs in the form of off-by-one errors and assertion
failures when the number of lines of code, number of div tags, and the
maximum line of a valid instruction were different.
This gets rid of all of that by
1. Only rendering stuff for which we have div tags (which excludes
everything not on screen)
2. If we have a missmatch between the number of lines of code and the
number of valid instructions, just take the smaller one.

Fix #11